+++ app/src/main/java/tv/withaibuild/customiuizer/MainFragment.kt
@@ -71,6 +71,7 @@ class MainFragment : PreferenceFragmentBase() {
         return act != null && !act.isFinishing && isAdded
     }
 
+    @SuppressLint("MissingSuperCall")
     override fun onCreate(savedInstanceState: Bundle?) {
         toolbarMenu = true
         activeMenus = "all"